% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_cluster_match.R
\name{commuting_zones}
\alias{commuting_zones}
\title{The Commuting Zones main function to get location clusters.}
\usage{
commuting_zones(data, location_col_name, country_col_name, gmaps_key = "")
}
\arguments{
\item{data}{dataset that holds locations that will be used.}

\item{location_col_name}{location column name in data in English to match lat
& long.}

\item{country_col_name}{country column name in data in English to make lat
long match easier.}

\item{gmaps_key}{string containing your Google Maps API key.}
}
\value{
A data frame where each row represents a location in location_data and their
specific cluster.
}
\description{
Match all location points to their clusters defined by user movement between
cities. The main wrapper function.
}
\examples{
\dontrun{
location_df <- data.frame(
  location = c("Austin", "Los Angeles", "Buenos Aires"),
  country = c("United States", "United States", "Argentina")
)

matched_df <- commuting_zones(
  data = location_df,
  location_col_name = "location",
  country_col_name = "country",
  gmaps_key = "Enter Your Google API Key Here"
)

head(matched_df$matched_spdf)
plot(matched_df)
}
}
