% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_cluster_match.R
\name{filter_cluster_file}
\alias{filter_cluster_file}
\title{Load Commuting Zones Clusters from Data For Good.}
\usage{
filter_cluster_file(country_name)
}
\arguments{
\item{country_name}{specific country in which clusters should be located.
Default is NULL.}
}
\value{
shapefile data.frame that holds values for region, country,
fbcz_id (cluster id), fbcz_id_num (cluster number), name, cz_gen_ds (date of
update), win_population, win_roads_km, area, geography (polygon with points
that form the cluster).
}
\description{
This method loads the Community Zones Clusters CSV into memory
and imports the file in a format compatible with the Shapefile library.
}
