% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_lat_long.R
\name{get_location_lat_long}
\alias{get_location_lat_long}
\title{Match location units to spatial location.}
\usage{
get_location_lat_long(
  data,
  location_col_name = "location",
  country_col_name = "country"
)
}
\arguments{
\item{data}{dataset that holds locations that will be used.}

\item{location_col_name}{location column name in data in English to match lat
& long.}

\item{country_col_name}{country column name in data in English to make lat
long match easier.}
}
\description{
Find location latitude and longitude using Gmaps API.
}
\examples{
\dontrun{
# Example dataset:
df <- data.frame(
  'location' = c("Austin", "Toronto"),
  'country' = c("United States", "Canada")
)
df <- get_location_lat_long(
  df,
  location_col_name = "location",
  country_col_name = "country"
)
}
}
