% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_cluster_match.R
\name{post_msgs}
\alias{post_msgs}
\title{Post messages letting user know how many locations were matched to their CZ
clusters.}
\usage{
post_msgs(df)
}
\arguments{
\item{df}{data.frame to run the analysis on.}
}
\description{
Post messages letting user know how many locations were matched to their CZ
clusters.
}
