% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zip_to_cz}
\alias{zip_to_cz}
\title{Match zip to their respective CZ cluster.}
\format{
A data frame with defined lat & long polygons for each cluster.
\describe{
   \item{zipcode}{Numeric Zip Code.}
   \item{region}{Continent.}
   \item{fbcz_id}{String: Commuting Zones id.}
   \item{fbcz_id_num}{Integer: unique Commuting Zones id.}
   \item{cz_gen_ds}{Date in which Commuting Zones were built.}
   \item{country_iso3}{ISO 3 code for country.}

}
}
\usage{
zip_to_cz
}
\description{
March 2023
}
\keyword{datasets}
