% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\docType{package}
\name{CommutingZones}
\alias{CommutingZones}
\alias{CommutingZones-package}
\title{CommutingZones}
\description{
Commuting zones are geographic areas where people live and work and are useful for understanding local economies, as well as how they differ from traditional boundaries. These zones are a set of boundary shapes built using aggregated estimates of home and work locations. Data used to build commuting zones is aggregated and de-identified.
}
\author{
\strong{Maintainer}: Monica Bhole \email{mbhole@meta.com}

Authors:
\itemize{
  \item Gufeng Zhou \email{gufeng@meta.com}
  \item Arturo Esquerra \email{aesquerrac@gmail.com}
  \item Nicolas Cruces \email{nicolascruces1994@gmail.com}
}

Other contributors:
\itemize{
  \item Igor Skokan \email{igorskokan@meta.com} [contributor]
}

}
