% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_lat_long.R
\name{get_location_lat_long}
\alias{get_location_lat_long}
\alias{check_gmaps_key}
\title{Match location units to spatial location.}
\usage{
get_location_lat_long(
  data,
  location_col_name = "location",
  country_col_name = "country",
  gmaps_key = NULL,
  quiet = FALSE,
  ...
)

check_gmaps_key(gmaps_key = NULL)
}
\arguments{
\item{data}{dataset that holds locations that will be used.}

\item{location_col_name}{location column name in data in English to match lat
& long.}

\item{country_col_name}{country column name in data in English to make lat
long match easier.}

\item{gmaps_key}{string containing your Google Maps API key.}

\item{quiet}{Boolean. Keep quiet? If not, show warnings and messages.}

\item{...}{Additional parameters}
}
\description{
Find location latitude and longitude using Gmaps API.
}
\examples{
\dontrun{
# Example dataset:
df <- data.frame(
  "location" = c("Austin", "Toronto"),
  "country" = c("United States", "Canada")
)
df <- get_location_lat_long(
  df,
  location_col_name = "location",
  country_col_name = "country",
  gmaps_key = NULL # Or change for valid key.
)
}
}
