% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_cluster_match.R
\name{location_to_cluster_match}
\alias{location_to_cluster_match}
\title{Match all locations to their cluster, via lat and long.}
\usage{
location_to_cluster_match(
  location_data,
  cluster_data,
  longitude_col_name = "longitude",
  latitude_col_name = "latitude"
)
}
\arguments{
\item{location_data}{data.frame object that holds the location name,
latitude and longitude for all locations.}

\item{cluster_data}{data.frame object with all polygons.  Each row represents
a different cluster with an sf::POLYGON or sf::MULTIPOLYGON object that has
the points that are joined to form that cluster.}

\item{longitude_col_name}{name of the longitude column in the location_data
data.frame.}

\item{latitude_col_name}{name of the latitude column in the location_data
data.frame.}
}
\value{
A data frame where each row represents a location in location_data and their
specific cluster.
}
\description{
Match all location points to their clusters defined by user movement between
cities.
}
\examples{
\dontrun{
location_df <- data.frame(
  location = c('Austin', 'New York'),
  country_name = c('United States', 'United States')
)
location_df <- get_location_lat_long(
  location_df,
  location_col_name='location',
  country_col_name='country_name')
cluster_file <- filter_cluster_file(country_name = 'United States')

matched_df <- location_to_cluster_match(
  location_df, cluster_file
}
}
